Fix data source internal resolution logic
When running end-to-end tests the data source may not be loaded and a hot reload is required. The existing logic would fail the retrieval of a datasource if the reference did not exist prior to the hot reload. Modified the logic so that the data source reference can be null and passing in the resource id for hot reload support.